+++ raillabel/format/bbox.py
@@ -67,7 +67,7 @@ class Bbox(_Annotation):
         # Adds the optional properties
         if "coordinate_system" in data_dict and data_dict["coordinate_system"] != "":
             try:
-                annotation.coordinate_system = coordinate_systems[data_dict["coordinate_system"]]
+                annotation.sensor = coordinate_systems[data_dict["coordinate_system"]]
 
             except KeyError:
                 warnings.append(