+++ raillabel/format/poly2d.py
@@ -85,7 +85,7 @@ class Poly2d(_Annotation):
         # Adds the optional properties
         if "coordinate_system" in data_dict and data_dict["coordinate_system"] != "":
             try:
-                annotation.coordinate_system = coordinate_systems[data_dict["coordinate_system"]]
+                annotation.sensor = coordinate_systems[data_dict["coordinate_system"]]
 
             except KeyError:
                 warnings.append(