+++ raillabel/format/scene.py
@@ -5,11 +5,10 @@ import typing as t
 import uuid
 from dataclasses import dataclass, field
 
-from .coordinate_system import CoordinateSystem
 from .frame import Frame
 from .metadata import Metadata
 from .object import Object
-from .stream import Stream
+from .sensor import Sensor
 
 
 @dataclass
@@ -20,11 +19,8 @@ class Scene:
     ----------
     metadata: raillabel.format.Metadata
         This object contains information, that is, metadata, about the annotation file itself.
-    streams: dict of raillabel.format.Stream, optional
-        Dictionary of raillabel.format.Streams. Dictionary keys are the stream uids. Default is {}.
-    coordinate_systems: dict of raillabel.format.CoordinateSystem, optional
-        Dictionary of raillabel.format.CoordinateSystems. Dictionary keys are the coordinate_system
-        uids. Default is {}.
+    sensors: dict of raillabel.format.Sensor, optional
+        Dictionary of raillabel.format.Sensors. Dictionary keys are the sensor uids. Default is {}.
     objects: dict of raillabel.format.Object, optional
         Dictionary of raillabel.format.Objects. Dictionary keys are the object uids. Default is {}.
     frames: dict of raillabel.format.Frame, optional
@@ -32,8 +28,7 @@ class Scene:
     """
 
     metadata: Metadata
-    streams: t.Dict[str, Stream] = field(default_factory=dict)
-    coordinate_systems: t.Dict[str, CoordinateSystem] = field(default_factory=dict)
+    sensors: t.Dict[str, Sensor] = field(default_factory=dict)
     objects: t.Dict[uuid.UUID, Object] = field(default_factory=dict)
     frames: t.Dict[int, Frame] = field(default_factory=dict)
 
@@ -53,14 +48,17 @@ class Scene:
 
         dict_repr = {"openlabel": {"metadata": self.metadata.asdict()}}
 
-        if self.streams != {}:
+        if self.sensors != {}:
             dict_repr["openlabel"]["streams"] = {
-                str(k): v.asdict() for k, v in self.streams.items()
+                str(k): v.asdict()["stream"] for k, v in self.sensors.items()
             }
-
-        if self.coordinate_systems != {}:
             dict_repr["openlabel"]["coordinate_systems"] = {
-                str(k): v.asdict() for k, v in self.coordinate_systems.items()
+                str(k): v.asdict()["coordinate_system"] for k, v in self.sensors.items()
+            }
+            dict_repr["openlabel"]["coordinate_systems"]["base"] = {
+                "type": "local",
+                "parent": "",
+                "children": list(self.sensors.keys()),
             }
 
         if self.objects != {}: