+++ tests/test_raillabel/format_loaders/test_loader_raillabel_v2.py
@@ -153,33 +153,33 @@ def test_load_frame0_metadata(openlabel_v1_short_data, loader):
     assert 0 in scene.frames
     assert scene.frames[0].timestamp == decimal.Decimal("1632321743.134149")
 
-    assert len(scene.frames[0].streams) == 3
-    assert "rgb_middle" in scene.frames[0].streams
-    assert scene.frames[0].streams["rgb_middle"].stream == scene.streams["rgb_middle"]
-    assert scene.frames[0].streams["rgb_middle"].timestamp == decimal.Decimal(
+    assert len(scene.frames[0].sensors) == 3
+    assert "rgb_middle" in scene.frames[0].sensors
+    assert scene.frames[0].sensors["rgb_middle"].sensor == scene.sensors["rgb_middle"]
+    assert scene.frames[0].sensors["rgb_middle"].timestamp == decimal.Decimal(
         "1632321743.100000072"
     )
-    assert scene.frames[0].streams["rgb_middle"].uri == "rgb_test0.png"
-    assert "ir_middle" in scene.frames[0].streams
-    assert scene.frames[0].streams["ir_middle"].stream == scene.streams["ir_middle"]
-    assert scene.frames[0].streams["ir_middle"].timestamp == decimal.Decimal("1632321743.106000004")
-    assert scene.frames[0].streams["ir_middle"].uri == "ir_test0.png"
-    assert "lidar" in scene.frames[0].streams
-    assert scene.frames[0].streams["lidar"].stream == scene.streams["lidar"]
-    assert scene.frames[0].streams["lidar"].timestamp == decimal.Decimal("1632321743.134149")
-    assert scene.frames[0].streams["lidar"].uri == "lidar_test0.pcd"
+    assert scene.frames[0].sensors["rgb_middle"].uri == "rgb_test0.png"
+    assert "ir_middle" in scene.frames[0].sensors
+    assert scene.frames[0].sensors["ir_middle"].sensor == scene.sensors["ir_middle"]
+    assert scene.frames[0].sensors["ir_middle"].timestamp == decimal.Decimal("1632321743.106000004")
+    assert scene.frames[0].sensors["ir_middle"].uri == "ir_test0.png"
+    assert "lidar" in scene.frames[0].sensors
+    assert scene.frames[0].sensors["lidar"].sensor == scene.sensors["lidar"]
+    assert scene.frames[0].sensors["lidar"].timestamp == decimal.Decimal("1632321743.134149")
+    assert scene.frames[0].sensors["lidar"].uri == "lidar_test0.pcd"
 
     assert len(scene.frames[0].data) == 2
     assert "test_frame_data0" in scene.frames[0].data
     assert scene.frames[0].data["test_frame_data0"].uid == "a06fe567-29c7-475b-92a4-fbca64e671a7"
     assert scene.frames[0].data["a06fe567-29c7-475b-92a4-fbca64e671a7"].name == "test_frame_data0"
     assert scene.frames[0].data["test_frame_data0"].val == 53.1
-    assert scene.frames[0].data["test_frame_data0"].sensor == scene.coordinate_systems["rgb_middle"]
+    assert scene.frames[0].data["test_frame_data0"].sensor == scene.sensors["rgb_middle"]
     assert "test_frame_data1" in scene.frames[0].data
     assert scene.frames[0].data["test_frame_data1"].uid == "4bb95df7-a051-48a9-b77e-72f27ca43f64"
     assert scene.frames[0].data["4bb95df7-a051-48a9-b77e-72f27ca43f64"].name == "test_frame_data1"
     assert scene.frames[0].data["test_frame_data1"].val == 10
-    assert scene.frames[0].data["test_frame_data1"].sensor == scene.coordinate_systems["lidar"]
+    assert scene.frames[0].data["test_frame_data1"].sensor == scene.sensors["lidar"]
 
     assert len(scene.frames[0].object_data) == 2
     assert "b40ba3ad-0327-46ff-9c28-2506cfd6d934" in scene.frames[0].object_data
@@ -256,7 +256,7 @@ def test_load_frame0_bboxs(openlabel_v1_short_data, loader):
         .object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"]
         .bboxs["78f0ad89-2750-4a30-9d66-44c9da73a714"]
         .sensor
-        == scene.coordinate_systems["rgb_middle"]
+        == scene.sensors["rgb_middle"]
     )
     assert scene.frames[0].object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"].bboxs[
         "78f0ad89-2750-4a30-9d66-44c9da73a714"
@@ -328,7 +328,7 @@ def test_load_frame0_bboxs(openlabel_v1_short_data, loader):
         .object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"]
         .bboxs["68b4e02c-40c8-4de0-89ad-bc00ed05a043"]
         .sensor
-        == scene.coordinate_systems["ir_middle"]
+        is scene.sensors["ir_middle"]
     )
     assert scene.frames[0].object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"].bboxs[
         "68b4e02c-40c8-4de0-89ad-bc00ed05a043"
@@ -398,7 +398,7 @@ def test_load_frame0_poly2ds(openlabel_v1_short_data, loader):
         .object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"]
         .poly2ds["bebfbae4-61a2-4758-993c-efa846b050a5"]
         .sensor
-        == scene.coordinate_systems["rgb_middle"]
+        == scene.sensors["rgb_middle"]
     )
     assert scene.frames[0].object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"].poly2ds[
         "bebfbae4-61a2-4758-993c-efa846b050a5"
@@ -477,7 +477,7 @@ def test_load_frame0_poly2ds(openlabel_v1_short_data, loader):
         .object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"]
         .poly2ds["3f63201c-fb33-4487-aff6-ae0aa5fa976c"]
         .sensor
-        == scene.coordinate_systems["ir_middle"]
+        == scene.sensors["ir_middle"]
     )
     assert scene.frames[0].object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"].poly2ds[
         "3f63201c-fb33-4487-aff6-ae0aa5fa976c"
@@ -588,7 +588,7 @@ def test_load_frame0_cuboids(openlabel_v1_short_data, loader):
         .object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"]
         .cuboids["dc2be700-8ee4-45c4-9256-920b5d55c917"]
         .sensor
-        == scene.coordinate_systems["lidar"]
+        == scene.sensors["lidar"]
     )
     assert scene.frames[0].object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"].cuboids[
         "dc2be700-8ee4-45c4-9256-920b5d55c917"
@@ -702,7 +702,7 @@ def test_load_frame0_cuboids(openlabel_v1_short_data, loader):
         .object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"]
         .cuboids["450ceb81-9778-4e63-bf89-42f3ed9f6747"]
         .sensor
-        == scene.coordinate_systems["lidar"]
+        == scene.sensors["lidar"]
     )
     assert scene.frames[0].object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"].cuboids[
         "450ceb81-9778-4e63-bf89-42f3ed9f6747"
@@ -768,7 +768,7 @@ def test_load_frame0_poly3ds(openlabel_v1_short_data, loader):
         .object_data["22dedd49-6dcb-413b-87ef-00ccfb532e98"]
         .poly3ds["14f58fb0-add7-4ed9-85b3-74615986d854"]
         .sensor
-        == scene.coordinate_systems["lidar"]
+        == scene.sensors["lidar"]
     )
 
 
@@ -809,7 +809,7 @@ def test_load_frame0_seg3ds(openlabel_v1_short_data, loader):
         .object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"]
         .seg3ds["c1087f1d-7271-4dee-83ad-519a4e3b78a8"]
         .sensor
-        == scene.coordinate_systems["lidar"]
+        == scene.sensors["lidar"]
     )
     assert scene.frames[0].object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"].seg3ds[
         "c1087f1d-7271-4dee-83ad-519a4e3b78a8"
@@ -856,7 +856,7 @@ def test_load_frame0_seg3ds(openlabel_v1_short_data, loader):
         .object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"]
         .seg3ds["50be7fe3-1f43-47ca-b65a-930e6cfacfeb"]
         .sensor
-        == scene.coordinate_systems["lidar"]
+        == scene.sensors["lidar"]
     )
     assert scene.frames[0].object_data["b40ba3ad-0327-46ff-9c28-2506cfd6d934"].seg3ds[
         "50be7fe3-1f43-47ca-b65a-930e6cfacfeb"
@@ -869,31 +869,31 @@ def test_load_frame1_metadata(openlabel_v1_short_data, loader):
     assert 1 in scene.frames
     assert scene.frames[1].timestamp == decimal.Decimal("1632321743.233263")
 
-    assert len(scene.frames[1].streams) == 3
-    assert "rgb_middle" in scene.frames[1].streams
-    assert scene.frames[1].streams["rgb_middle"].stream == scene.streams["rgb_middle"]
-    assert scene.frames[1].streams["rgb_middle"].timestamp == decimal.Decimal("1632321743.2")
-    assert scene.frames[1].streams["rgb_middle"].uri == "rgb_test1.png"
-    assert "ir_middle" in scene.frames[1].streams
-    assert scene.frames[1].streams["ir_middle"].stream == scene.streams["ir_middle"]
-    assert scene.frames[1].streams["ir_middle"].timestamp == decimal.Decimal("1632321743.208000004")
-    assert scene.frames[1].streams["ir_middle"].uri == "ir_test1.png"
-    assert "lidar" in scene.frames[1].streams
-    assert scene.frames[1].streams["lidar"].stream == scene.streams["lidar"]
-    assert scene.frames[1].streams["lidar"].timestamp == decimal.Decimal("1632321743.233263")
-    assert scene.frames[1].streams["lidar"].uri == "lidar_test1.pcd"
+    assert len(scene.frames[1].sensors) == 3
+    assert "rgb_middle" in scene.frames[1].sensors
+    assert scene.frames[1].sensors["rgb_middle"].sensor == scene.sensors["rgb_middle"]
+    assert scene.frames[1].sensors["rgb_middle"].timestamp == decimal.Decimal("1632321743.2")
+    assert scene.frames[1].sensors["rgb_middle"].uri == "rgb_test1.png"
+    assert "ir_middle" in scene.frames[1].sensors
+    assert scene.frames[1].sensors["ir_middle"].sensor == scene.sensors["ir_middle"]
+    assert scene.frames[1].sensors["ir_middle"].timestamp == decimal.Decimal("1632321743.208000004")
+    assert scene.frames[1].sensors["ir_middle"].uri == "ir_test1.png"
+    assert "lidar" in scene.frames[1].sensors
+    assert scene.frames[1].sensors["lidar"].sensor == scene.sensors["lidar"]
+    assert scene.frames[1].sensors["lidar"].timestamp == decimal.Decimal("1632321743.233263")
+    assert scene.frames[1].sensors["lidar"].uri == "lidar_test1.pcd"
 
     assert len(scene.frames[1].data) == 2
     assert "test_frame_data0" in scene.frames[1].data
     assert scene.frames[1].data["test_frame_data0"].uid == "558697df-61f5-41b0-b112-3d6fcbd7d6c9"
     assert scene.frames[1].data["558697df-61f5-41b0-b112-3d6fcbd7d6c9"].name == "test_frame_data0"
     assert scene.frames[1].data["test_frame_data0"].val == 53
-    assert scene.frames[1].data["test_frame_data0"].sensor == scene.coordinate_systems["rgb_middle"]
+    assert scene.frames[1].data["test_frame_data0"].sensor == scene.sensors["rgb_middle"]
     assert "test_frame_data1" in scene.frames[1].data
     assert scene.frames[1].data["test_frame_data1"].uid == "843e07a0-aac5-4f62-8200-1468dbd3055d"
     assert scene.frames[1].data["843e07a0-aac5-4f62-8200-1468dbd3055d"].name == "test_frame_data1"
     assert scene.frames[1].data["test_frame_data1"].val == 10.1
-    assert scene.frames[1].data["test_frame_data1"].sensor == scene.coordinate_systems["lidar"]
+    assert scene.frames[1].data["test_frame_data1"].sensor == scene.sensors["lidar"]
 
     assert len(scene.frames[1].object_data) == 2
     assert "6fe55546-0dd7-4e40-b6b4-bb7ea3445772" in scene.frames[1].object_data
@@ -970,7 +970,7 @@ def test_load_frame1_bboxs(openlabel_v1_short_data, loader):
         .object_data["6fe55546-0dd7-4e40-b6b4-bb7ea3445772"]
         .bboxs["6ba42cbc-484e-4b8d-a022-b23c2bb6643c"]
         .sensor
-        == scene.coordinate_systems["rgb_middle"]
+        == scene.sensors["rgb_middle"]
     )
     assert scene.frames[1].object_data["6fe55546-0dd7-4e40-b6b4-bb7ea3445772"].bboxs[
         "6ba42cbc-484e-4b8d-a022-b23c2bb6643c"
@@ -1042,7 +1042,7 @@ def test_load_frame1_bboxs(openlabel_v1_short_data, loader):
         .object_data["6fe55546-0dd7-4e40-b6b4-bb7ea3445772"]
         .bboxs["5f28fa18-8f2a-4a40-a0b6-c0bbedc00f2e"]
         .sensor
-        == scene.coordinate_systems["ir_middle"]
+        == scene.sensors["ir_middle"]
     )
     assert scene.frames[1].object_data["6fe55546-0dd7-4e40-b6b4-bb7ea3445772"].bboxs[
         "5f28fa18-8f2a-4a40-a0b6-c0bbedc00f2e"
@@ -1112,7 +1112,7 @@ def test_load_frame1_poly2ds(openlabel_v1_short_data, loader):
         .object_data["6fe55546-0dd7-4e40-b6b4-bb7ea3445772"]
         .poly2ds["e2503c5d-9fe4-4666-b510-ef644c5a766b"]
         .sensor
-        == scene.coordinate_systems["rgb_middle"]
+        == scene.sensors["rgb_middle"]
     )
     assert scene.frames[1].object_data["6fe55546-0dd7-4e40-b6b4-bb7ea3445772"].poly2ds[
         "e2503c5d-9fe4-4666-b510-ef644c5a766b"
@@ -1232,7 +1232,7 @@ def test_load_frame1_cuboids(openlabel_v1_short_data, loader):
         .object_data["22dedd49-6dcb-413b-87ef-00ccfb532e98"]
         .cuboids["536ac83a-32c8-4fce-8499-ef32716c64a6"]
         .sensor
-        == scene.coordinate_systems["lidar"]
+        == scene.sensors["lidar"]
     )
     assert scene.frames[1].object_data["22dedd49-6dcb-413b-87ef-00ccfb532e98"].cuboids[
         "536ac83a-32c8-4fce-8499-ef32716c64a6"
@@ -1346,7 +1346,7 @@ def test_load_frame1_cuboids(openlabel_v1_short_data, loader):
         .object_data["22dedd49-6dcb-413b-87ef-00ccfb532e98"]
         .cuboids["e53bd5e3-980a-4fa7-a0f9-5a2e59ba663c"]
         .sensor
-        == scene.coordinate_systems["lidar"]
+        == scene.sensors["lidar"]
     )
     assert scene.frames[1].object_data["22dedd49-6dcb-413b-87ef-00ccfb532e98"].cuboids[
         "e53bd5e3-980a-4fa7-a0f9-5a2e59ba663c"
@@ -1390,7 +1390,7 @@ def test_load_frame1_seg3ds(openlabel_v1_short_data, loader):
         .object_data["22dedd49-6dcb-413b-87ef-00ccfb532e98"]
         .seg3ds["550df2c3-0e66-483e-bcc6-f3013b7e581b"]
         .sensor
-        == scene.coordinate_systems["lidar"]
+        == scene.sensors["lidar"]
     )
     assert scene.frames[1].object_data["22dedd49-6dcb-413b-87ef-00ccfb532e98"].seg3ds[
         "550df2c3-0e66-483e-bcc6-f3013b7e581b"
@@ -1437,7 +1437,7 @@ def test_load_frame1_seg3ds(openlabel_v1_short_data, loader):
         .object_data["22dedd49-6dcb-413b-87ef-00ccfb532e98"]
         .seg3ds["12b21c52-06ea-4269-9805-e7167e7a74ed"]
         .sensor
-        == scene.coordinate_systems["lidar"]
+        == scene.sensors["lidar"]
     )
     assert scene.frames[1].object_data["22dedd49-6dcb-413b-87ef-00ccfb532e98"].seg3ds[
         "12b21c52-06ea-4269-9805-e7167e7a74ed"
@@ -1472,36 +1472,27 @@ def test_load_uri_vcd_incompatible(
     assert scene == scene_ground_truth
 
 
-# Tests the warnings
+# Tests the warnings and errors
 def test_no_warnings(openlabel_v1_short_data, loader):
     loader.load(openlabel_v1_short_data, validate=False)
     assert len(loader.warnings) == 0
 
 
-def test_warnings_cs_parent(openlabel_v1_short_data, loader):
-    openlabel_v1_short_data["openlabel"]["coordinate_systems"]["lidar"]["parent"] = "test_cs"
+def test_stream_with_no_coordinate_system(openlabel_v1_short_data, loader):
+    del openlabel_v1_short_data["openlabel"]["coordinate_systems"]["ir_middle"]
+    del openlabel_v1_short_data["openlabel"]["coordinate_systems"]["base"]["children"][
+        openlabel_v1_short_data["openlabel"]["coordinate_systems"]["base"]["children"].index(
+            "ir_middle"
+        )
+    ]
+    scene = loader.load(openlabel_v1_short_data)
 
-    loader.load(openlabel_v1_short_data, validate=False)
     assert len(loader.warnings) == 1
-
-    # Tests for keywords in the warning that can help the user identify the source
+    assert "ir_middle" in loader.warnings[0]
+    assert "stream" in loader.warnings[0].lower()
     assert "coordinate system" in loader.warnings[0]
-    assert "parent" in loader.warnings[0]
-    assert "lidar" in loader.warnings[0]
-    assert "test_cs" in loader.warnings[0]
 
-
-def test_warnings_cs_child(openlabel_v1_short_data, loader):
-    openlabel_v1_short_data["openlabel"]["coordinate_systems"]["base"]["children"].append("test_cs")
-
-    loader.load(openlabel_v1_short_data, validate=False)
-    assert len(loader.warnings) == 1
-
-    # Tests for keywords in the warning that can help the user identify the source
-    assert "coordinate system" in loader.warnings[0]
-    assert "child" in loader.warnings[0]
-    assert "base" in loader.warnings[0]
-    assert "test_cs" in loader.warnings[0]
+    assert scene.sensors["ir_middle"].extrinsics is None
 
 
 def test_warnings_sync(openlabel_v1_short_data, loader):
@@ -1516,7 +1507,7 @@ def test_warnings_sync(openlabel_v1_short_data, loader):
     assert "frame" in loader.warnings[0]
     assert "0" in loader.warnings[0]
     assert "sync" in loader.warnings[0]
-    assert "test_stream" in loader.warnings[0]
+    assert "non_existing_stream" in loader.warnings[0]
 
 
 def test_warnings_stream_sync_field(openlabel_v1_short_data, loader):